[AND-146] Move ui logic from updates to the presentation package
